Make python3's getrandom() call non-blocking
during the first seconds of boot we may not have random entropy,
and the getrandom() system call will block at that point, stopping
python execution.

This added patch passes the flag to the getrandom() syscall to
make the call non-blocking. We may not get all the entropy we wanted,
but the code is robust for that (it will retry, or even fall back
to /dev/urandom)