Fix installation of base package with DNF
The base package was not being created, and DNF is not able to resolve the
package name 'python3'. This makes sense. Prior to this point in time, Yum has
somehow been able to resolve the name...

Fix the issue by properly instantiating the base package, with appropriate
runtime dependencies on -core, -lib, and -lib-avx2 subpackages.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>