python3: provide traceback for error reporting
It may be useful to report some unhandled Python exceptions via telemetry.
In particular, we are interested in exceptions in Python code originated in
/usr/bin, as these would be most likely resulting from files distributed by Clear OS.

Implementation:
The goal is to implemet the least invasive code change possible. We patch the Python code
(single file, single routine) that prints out tracebacks of unhadled exceptions.

If telemetry is enabled (and we are not in Python interactive mode), we intercept stderr
and  pipe the error text (traceback) into a temporary file. At the same time we also
output all messages via the original stderr as well so the user receives all messages
as usual. (The original sys.excepthook stays unmodified, allowing us to report
any exceptions in sys.excepthook as well, along with the original exception.)

Once we captured the traceback in a temporary file, we validate if the traceback
originated in /usr/bin. If so, we move (via "rename") the traceback file into a location
(/var/lib/telemetry/python) created and monitored by a python telemetry probe.
The probe will detect the new file and send the traceback to the telemetry backend
as a payload. The python probe itself is a installed as a part of telemetry.

If telemetry is not enabled, we do nothing. If we encounter any errors we do not
create the telemetry file.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>