0001-pythonrun.c-telemetry-patch: modified reporting
Instead of storing the tracback messages in a file in a "well known"
location and waiting for a python probe to pick it up and report to the
back-end, we send the traceback file directly to the back-end.
We use the telemetry library /usr/lib64/libtelemetry.so.3 if
present. The library is only present if user installed telemetry bundle.
If the library is not present, we don't report anything.
It is the responibility of the library not to report anything
if telemetry is installed but the user is not opted in.

This change removes the need for any python-probe services and creation
of "well known" directories.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>