Add python usrlocal.pth file
Add a path file to add /usr/local/ to the sys.path to allow users to
install system python packages that aren't in Clear Linux in a
way that doesn't potentially conflict with Clear Linux content.

Also cleanup the Provides from the package to avoid duplication and
conflicts that cause difficulty with dependency resolution.