Fix python build of distutils
Python can't use the math module when building itself so an
alternative to avoid the previous file based race condition must be
used. In addition to just using the return code of os.system to
validate if the library could be found.

Also only evaluate each library in isolation to avoid cases where an
earlier library that doesn't exist in the path causing the compilation
to fail.