Revert library flags change
This was done in order to get perf to build correctly but the change
ended up being done on the perf end instead since python-config needed
to be called differently to get the result needed. With that change to
the perf build process things work as expected so this change can be
removed.