Don't provide typing_extensions
typing_extensions is too widespread to avoid packaging as it isn't
conditionally checked for in many packages. In cases where it isn't
conditionally checked for, the potential of different behavior between
the module and the stdlib makes this not something we can easily patch
ourselves out of using either.

For now accept typing_extensions as part of the module ecosystem.

Signed-off-by: William Douglas <william.douglas@intel.com>