Fix build of tkinter extension
Starting with python 3.11, the extension module dependencies are queried
with pkg-config.

A side effect of this change is that additional buildreqs need to be
present to fix pkg-config checks for tcl/tk, since they reference static
libraries, and the tk8*.so is in tk-extras.

Add those buildreqs to enable/fix the tkinter build.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>