Rework python build
There is no longer a need to split the optimized confgure build from
shared as they build fine together as of 3.12 at least.

This fixes a number of issues with how python installed and so the
rest of this change cleans up the hacks those issues presented for
providing shared libraries and static libraries.

Signed-off-by: William Douglas <william.douglas@intel.com>