Reorder %build and fix %files
The non-optimized build should go last as they are in subdirectories
that won't as likely pollute the main build (the main build seems to
cause the optimization run to fail for subsequent builds).

Also fixup files for libpython3.so in VA and reorder contents to be a
little easier to find.

Signed-off-by: William Douglas <william.douglas@intel.com>