Move back to python being primarily staticly built
This change reverts back to using a staticly built python for the most
part. Rather than the previous method of mix building and installing
the dynamic python, this chnage has a one off build for just the
dynamic library. The library is manually installed and it is only
built for the base arch (the dynamic library can have issues if build
and runtime aren't exactly the same).

Signed-off-by: William Douglas <william.douglas@intel.com>